+++ Jogo_COO/src/GameLib/GameLib.java
@@ -1,3 +1,5 @@
+package GameLib;
+
 import java.awt.Color;
 import java.awt.Graphics;
 import java.awt.Toolkit;
