+++ Jogo_COO/src/GameObjects/Colliders/CollideWithPlayer.java
@@ -1,0 +1,7 @@
+package GameObjects.Colliders;
+
+import GameObjects.SpaceShips.Player;
+
+public interface CollideWithPlayer {
+    void colideWithPlayer(Player player);
+}
