+++ Jogo_COO/src/GameObjects/PowerUps/PowerUp.java
@@ -1,0 +1,7 @@
+package GameObjects.PowerUps;
+
+import GameObjects.GameObject;
+
+public abstract class PowerUp extends GameObject{
+
+}
