+++ Jogo_COO/src/GameObjects/SpaceShips/Enemies/Bosses/Boss.java
@@ -1,0 +1,5 @@
+package GameObjects.SpaceShips.Enemies.Bosses;
+
+public class Boss {
+    
+}
