+++ Jogo_COO/src/GameObjects/SpaceShips/Enemies/Enemy1.java
@@ -1,0 +1,47 @@
+package GameObjects.SpaceShips.Enemies;
+
+import java.awt.Color;
+import java.util.List;
+import java.util.Random;
+import GameObjects.Colliders.*;
+import GameLib.GameLib;
+import GameObjects.Projectiles.*;
+
+
+public class Enemy1 extends Enemy{
+        // construtor
+        public Enemy1(double x, double y, double escalarVelocity, double angle, double velocityRotation){
+            super(1, x, y, escalarVelocity, angle, velocityRotation, ENEMY1_RADIUS, 1.0, 1);
+            color = Color.CYAN; // cor do inimigo tipo 1
+        }
+
+        // métodos
+        public void drawShape(){
+            if(isStateTrue(ACTIVE)){
+                GameLib.drawCircle(getX(), getY(), ENEMY1_RADIUS);
+            }
+        }
+
+        public void moveAndDirection(long time){
+            x += getEscalarVelocity() * Math.cos(getAngle()) * time;
+            y += getEscalarVelocity() * Math.sin(getAngle()) * time * (-1.0);
+            angle += getRotationVelocity() * time;
+        }
+
+        public void shoot(List<EnemyProjectile> enemyProjectiles, List<CollideWithPlayer> colideComPlayer) {
+            // disparo de projetil
+            if(System.currentTimeMillis() > getNextShoot()){
+                Random rand = new Random();
+                EnemyProjectile proj = new EnemyProjectile(
+                    getX(),
+                    getY(),
+                    Math.cos(getAngle()) * 0.45,
+                    Math.sin(getAngle()) * 0.45 * (-1.0)
+                );
+                
+                enemyProjectiles.add(proj);
+                colideComPlayer.add(proj);
+                setNextShoot((long)(System.currentTimeMillis() + 200 + rand.nextDouble() * 500));
+            }
+        }
+    }
